Reduce code duplication for stepping DSL tests
